+++ frontend/src/stories/Button.stories.ts
@@ -1,32 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/web-components";
 import { html } from "lit";
-import { ifDefined } from "lit/directives/if-defined.js";
 
-import type { Button as ButtonComponent } from "@/components/ui/button";
-
-import "@/components/ui/button";
-
-const renderButton = ({
-  variant,
-  filled,
-  label,
-  raised,
-  loading,
-  href,
-}: ButtonComponent) => {
-  return html`
-    <btrix-button
-      variant=${ifDefined(variant)}
-      label=${ifDefined(label)}
-      href=${ifDefined(href)}
-      ?filled=${filled}
-      ?raised=${raised}
-      ?loading=${loading}
-    >
-      ${label}
-    </btrix-button>
-  `;
-};
+import { renderButton, RenderProps } from "./Button";
 
 const meta = {
   component: "btrix-button",
@@ -35,19 +10,15 @@ const meta = {
   argTypes: {
     type: {
       control: { type: "select" },
-      options: ["button", "submit"] satisfies ButtonComponent["type"][],
+      options: ["button", "submit"] satisfies RenderProps["type"][],
     },
     variant: {
       control: { type: "select" },
-      options: ["neutral", "danger"] satisfies ButtonComponent["variant"][],
+      options: ["neutral", "danger"] satisfies RenderProps["variant"][],
     },
     size: {
       control: { type: "select" },
-      options: [
-        "x-small",
-        "small",
-        "medium",
-      ] satisfies ButtonComponent["size"][],
+      options: ["x-small", "small", "medium"] satisfies RenderProps["size"][],
     },
   },
   args: {
@@ -59,10 +30,10 @@ const meta = {
       showPanel: false,
     },
   },
-} satisfies Meta<ButtonComponent>;
+} satisfies Meta<RenderProps>;
 
 export default meta;
-type Story = StoryObj<ButtonComponent>;
+type Story = StoryObj<RenderProps>;
 
 export const Raised: Story = {
   args: {
@@ -78,16 +49,32 @@ export const Loading: Story = {
 
 export const Variants: Story = {
   render: () => html`
-    <btrix-button filled variant="neutral">Neutral (Default)</btrix-button>
-    <btrix-button filled variant="danger">Danger</btrix-button>
+    ${renderButton({
+      variant: "neutral",
+      label: "Neutral (Default)",
+      filled: true,
+    })}
+    ${renderButton({ variant: "danger", label: "Danger", filled: true })}
   `,
 };
 
 export const Sizes: Story = {
   render: () => html`
-    <btrix-button filled size="x-small">X-Small</btrix-button>
-    <btrix-button filled size="small">Small</btrix-button>
-    <btrix-button filled size="medium">Medium (Default)</btrix-button>
+    ${renderButton({
+      size: "x-small",
+      label: "X-Small",
+      filled: true,
+    })}
+    ${renderButton({
+      size: "small",
+      label: "Small",
+      filled: true,
+    })}
+    ${renderButton({
+      size: "medium",
+      label: "Medium (Default",
+      filled: true,
+    })}
   `,
 };
